requirements.txt
chart_studio==1.1.0
dash==2.1.0
dash_bootstrap_components==1.0.3
dash_core_components==2.0.0
dash_html_components==2.0.0
dash_renderer==1.9.1
dash_table==5.0.0
matplotlib==3.4.3
pandas==1.3.4
plotly==5.5.0
PyYAML==6.0
scikit_learn==1.0.2
seaborn==0.11.2
urllib3==1.26.7
Flask==2.1.0
jinja2<3.1.0
stripe==2.56.0
python-dotenv==0.19.0